refactor: transition to ORM by removing legacy SQL dependencies and updating documentation
