+++ .env.template
@@ -9,7 +9,7 @@ FMP_API_KEY=your_api_key_here
 DATABASE_URL=
 PGHOST=your_db_host
 PGPORT=5432
-PGDATABASE=market_data
+PGDATABASE=your_db_name
 PGUSER=your_username
 PGPASSWORD=your_password
 
@@ -28,9 +28,9 @@ MARKET_CLOSE=21:00
 CSV_PATH=
 
 # Cron schedules
-COLLECTION_SCHEDULE=0 * * * *
-STG_TO_CORE_SCHEDULE=0 2 * * *
-BACKUP_CRON_SCHEDULE=0 2 * * *
+COLLECTION_SCHEDULE="0 * * * *"
+STG_TO_CORE_SCHEDULE="0 2 * * *"
+BACKUP_CRON_SCHEDULE="0 2 * * *"
 
 # Cron installer behavior for setup_cronjob_*.sh
 # user: project-local wrapper + user crontab (default)
@@ -53,6 +53,6 @@ SLOT_NAME=replica_slot_1
 DATA_DIR=/var/lib/pgsql/16/data
 
 # Backup configuration (for setup_scripts/database_setup/*)
-BACKUP_DIR=/var/backups/postgres
+BACKUP_DIR=/var/lib/pgsql/16/backups
 DB_USER=postgres
 DB_PASSWORD=change_me
