+++ alembic.ini
@@ -23,7 +23,7 @@ qualname = sqlalchemy.engine
 
 [logger_alembic]
 level = INFO
-handlers = console
+handlers =
 qualname = alembic
 
 [handler_console]
