+++ alembic/versions/20260312_0001_baseline_schema.py
@@ -78,11 +78,11 @@ def upgrade() -> None:
         sa.Column("market_data_id", sa.BigInteger(), autoincrement=True, nullable=False),
         sa.Column("symbol", sa.Text(), nullable=False),
         sa.Column("ts", sa.DateTime(timezone=True), nullable=False),
-        sa.Column("open", sa.Numeric(18, 6), nullable=False),
-        sa.Column("high", sa.Numeric(18, 6), nullable=False),
-        sa.Column("low", sa.Numeric(18, 6), nullable=False),
-        sa.Column("close", sa.Numeric(18, 6), nullable=False),
-        sa.Column("volume", sa.BigInteger(), nullable=False),
+        sa.Column("open", sa.Numeric(18, 6), nullable=True),
+        sa.Column("high", sa.Numeric(18, 6), nullable=True),
+        sa.Column("low", sa.Numeric(18, 6), nullable=True),
+        sa.Column("close", sa.Numeric(18, 6), nullable=True),
+        sa.Column("volume", sa.BigInteger(), nullable=True),
         sa.Column("asset_type", sa.Text(), nullable=False),
         sa.Column("source", sa.Text(), nullable=True),
         sa.Column("created_at", sa.DateTime(timezone=True), server_default=sa.text("now()"), nullable=True),
